Refactor IntroActivity.kt
