+++ app/src/main/java/com/ggu/avd/workers/DrawableDatabaseWorker.kt
@@ -3,7 +3,9 @@ package com.ggu.avd.workers
 import android.content.Context
 import android.util.Log
 import androidx.work.CoroutineWorker
+import androidx.work.Data
 import androidx.work.WorkerParameters
+import androidx.work.workDataOf
 import com.ggu.avd.data.AppDatabase
 import com.ggu.avd.data.AvdDrawable
 import com.ggu.avd.utilities.DRAWABLE_DATA_FILENAME
@@ -12,10 +14,13 @@ import com.google.gson.reflect.TypeToken
 import com.google.gson.stream.JsonReader
 import kotlinx.coroutines.coroutineScope
 
+const val KEY_RESULT = "result"
+
 class DrawableDatabaseWorker(
     context: Context,
     workerParameters: WorkerParameters
 ) : CoroutineWorker(context, workerParameters){
+
     override suspend fun doWork(): Result = coroutineScope {
         try{
             applicationContext.assets.open(DRAWABLE_DATA_FILENAME).use { inputStream ->
@@ -23,17 +28,23 @@ class DrawableDatabaseWorker(
                     val drawableType = object : TypeToken<List<AvdDrawable>>(){}.type
                     val drawableList : List<AvdDrawable> = Gson().fromJson(jsonReader, drawableType)
 
-                    val database = AppDatabase.getInstance(applicationContext)
-                    database.drawableDao().insertAll(drawableList)
+                    val insert = insertAll(drawableList)
 
-                    Result.success()
+                    val output:Data = workDataOf(KEY_RESULT to insert)
+
+                    Result.success(output)
                 }
             }
-            Result.success()
         } catch (ex:Exception){
             Log.e(TAG, "Error drawable database", ex)
             Result.failure()
         }
+    }
+
+    private suspend fun insertAll(drawableList:List<AvdDrawable>){
+        val database = AppDatabase.getInstance(applicationContext)
+        database.drawableDao().insertAll(drawableList)
+
     }
 
     companion object {
